Refactor repo structure: move docs, static assets, and config into proper packages
- Move documentation files (ARCHITECTURE.md, CLI.md, EVAL.md, EXTENDING.md, ML.md, SETUP.md) to docs/
- Move static assets (favicons, logo.svg) to static/
- Move config.py to vtsearch/config.py
- Extract clip initialization and embedding cache logic from app.py into vtsearch/clips.py
- Update all import references across the codebase
- Slim down app.py to focus on Flask setup and CLI argument parsing
