+++ tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-import config
+import vtsearch.config as config
 
 # Reduce training epochs for faster tests (default is 200; 30 is sufficient
 # for the tiny MLP to converge on the small test dataset).
@@ -9,7 +9,7 @@ config.TRAIN_EPOCHS = 30
 import app as app_module
 
 # Import refactored modules and make them accessible through app_module
-from config import NUM_CLIPS, SAMPLE_RATE
+from vtsearch.config import NUM_CLIPS, SAMPLE_RATE
 from vtsearch.audio import generate_wav
 from vtsearch.models import initialize_models, train_and_score
 from vtsearch.models.progress import clear_progress_cache
