+++ tests/test_clips.py
@@ -41,7 +41,7 @@ class TestInitClips:
 
     def test_deterministic_embeddings(self):
         """CLAP embeddings should be deterministic for the same input audio."""
-        from config import DATA_DIR
+        from vtsearch.config import DATA_DIR
         from vtsearch.models import embed_audio_file
 
         clip = app_module.clips[1]
